+++ test/cbc_tests.jl
@@ -20,8 +20,8 @@ if @isdefined Cbc
     @info "Testing scenario solution with Cbc."
 
     @testset "Solving storage_test with Cbc" begin
-        dbfile = joinpath(@__DIR__, "storage_test.sqlite")
-        #dbfile = "c:/temp/storage_test.sqlite"
+        dbfile = joinpath(dbfile_path, "storage_test.sqlite")
+
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         # Test with default outputs
@@ -195,8 +195,7 @@ if @isdefined Cbc
     end  # "Solving storage_test with Cbc"
 
     @testset "Solving storage_transmission_test with Cbc" begin
-        dbfile = joinpath(@__DIR__, "storage_transmission_test.sqlite")
-        #dbfile = "c:/temp/storage_transmission_test.sqlite"
+        dbfile = joinpath(dbfile_path, "storage_transmission_test.sqlite")
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         @info "Running Cbc test 1 on storage_transmission_test.sqlite: default outputs."
@@ -242,8 +241,7 @@ if @isdefined Cbc
     end  # "Solving storage_transmission_test with Cbc"
 
     @testset "Solving ramp_test with Cbc" begin
-        dbfile = joinpath(@__DIR__, "ramp_test.sqlite")
-        #dbfile = "c:/temp/ramp_test.sqlite"
+        dbfile = joinpath(dbfile_path, "ramp_test.sqlite")
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         @info "Running Cbc test 1 on ramp_test.sqlite: default outputs."