+++ test/gurobi_tests.jl
@@ -20,8 +20,7 @@ if @isdefined Gurobi
     @info "Testing scenario solution with Gurobi."
 
     @testset "Solving storage_test with Gurobi" begin
-        dbfile = joinpath(@__DIR__, "storage_test.sqlite")
-        #dbfile = "c:/temp/storage_test.sqlite"
+        dbfile = joinpath(dbfile_path, "storage_test.sqlite")
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         # Test with default outputs
@@ -191,8 +190,7 @@ if @isdefined Gurobi
     end  # "Solving storage_test with Gurobi"
 
     @testset "Solving storage_transmission_test with Gurobi" begin
-        dbfile = joinpath(@__DIR__, "storage_transmission_test.sqlite")
-        #dbfile = "c:/temp/storage_transmission_test.sqlite"
+        dbfile = joinpath(dbfile_path, "storage_transmission_test.sqlite")
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         @info "Running Gurobi test 1 on storage_transmission_test.sqlite: default outputs."
@@ -316,8 +314,7 @@ if @isdefined Gurobi
     end  # "Solving storage_transmission_test with Gurobi"
 
     @testset "Solving ramp_test with Gurobi" begin
-        dbfile = joinpath(@__DIR__, "ramp_test.sqlite")
-        #dbfile = "c:/temp/ramp_test.sqlite"
+        dbfile = joinpath(dbfile_path, "ramp_test.sqlite")
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         @info "Running Gurobi test 1 on ramp_test.sqlite: default outputs."