+++ test/mosek_tests.jl
@@ -21,8 +21,7 @@ if @isdefined MosekTools
     @info "Testing scenario solution with Mosek."
 
     @testset "Solving storage_test with Mosek" begin
-        dbfile = joinpath(@__DIR__, "storage_test.sqlite")
-        #dbfile = "c:/temp/storage_test.sqlite"
+        dbfile = joinpath(dbfile_path, "storage_test.sqlite")
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         # Test with default outputs
@@ -192,8 +191,7 @@ if @isdefined MosekTools
     end  # "Solving storage_test with Mosek"
 
     @testset "Solving storage_transmission_test with Mosek" begin
-        dbfile = joinpath(@__DIR__, "storage_transmission_test.sqlite")
-        #dbfile = "c:/temp/storage_transmission_test.sqlite"
+        dbfile = joinpath(dbfile_path, "storage_transmission_test.sqlite")
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         @info "Running Mosek test 1 on storage_transmission_test.sqlite: default outputs."
@@ -306,8 +304,7 @@ if @isdefined MosekTools
     end  # "Solving storage_transmission_test with Mosek"
 
     @testset "Solving ramp_test with Mosek" begin
-        dbfile = joinpath(@__DIR__, "ramp_test.sqlite")
-        #dbfile = "c:/temp/ramp_test.sqlite"
+        dbfile = joinpath(dbfile_path, "ramp_test.sqlite")
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         @info "Running Mosek test 1 on ramp_test.sqlite: default outputs."