+++ test/xpress_tests.jl
@@ -20,8 +20,7 @@ if @isdefined Xpress
     @info "Testing scenario solution with Xpress."
 
     @testset "Solving storage_test with Xpress" begin
-        dbfile = joinpath(@__DIR__, "storage_test.sqlite")
-        #dbfile = "c:/temp/storage_test.sqlite"
+        dbfile = joinpath(dbfile_path, "storage_test.sqlite")
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         # Test with default outputs
@@ -192,8 +191,7 @@ if @isdefined Xpress
     end  # "Solving storage_test with Xpress"
 
     @testset "Solving storage_transmission_test with Xpress" begin
-        dbfile = joinpath(@__DIR__, "storage_transmission_test.sqlite")
-        #dbfile = "c:/temp/storage_transmission_test.sqlite"
+        dbfile = joinpath(dbfile_path, "storage_transmission_test.sqlite")
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         @info "Running Xpress test 1 on storage_transmission_test.sqlite: default outputs."
@@ -315,8 +313,7 @@ if @isdefined Xpress
     end  # "Solving storage_transmission_test with Xpress"
 
     @testset "Solving ramp_test with Xpress" begin
-        dbfile = joinpath(@__DIR__, "ramp_test.sqlite")
-        #dbfile = "c:/temp/ramp_test.sqlite"
+        dbfile = joinpath(dbfile_path, "ramp_test.sqlite")
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         @info "Running Xpress test 1 on ramp_test.sqlite: default outputs."